Add component for displaying key value data
